[breaking] gRPC Compile request now returns expanded build properties by default
