+++ rpc/cc/arduino/cli/commands/v1/compile.proto
@@ -91,6 +91,9 @@ message CompileRequest {
   // very specific reason to do so. This flag is mainly provided for usage in
   // language servers to optimize the build speed in some particular cases.
   bool skip_libraries_discovery = 28;
+  // If set to true the returned build properties will be left unexpanded, with
+  // the variables placeholders exactly as defined in the platform.
+  bool do_not_expand_build_properties = 29;
 }
 
 message CompileResponse {
